requirements.txt
Django==2.2.2
django-excel==0.0.10
django-rest-framework==0.1.0
djangorestframework==3.9.4
lml==0.0.9
numpy==1.16.4
pandas==0.24.2
pyexcel==0.5.14
pyexcel-io==0.5.18
pyexcel-webio==0.1.4
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
texttable==1.6.1
xlrd==1.2.0
